genimage: move extended partition to slot 3
Without the rescue partition, rootfs.0 ends up
in partition 3 and rootfs.1 in partition 5 (so
the mapping to partitions 4 and 5 was wrong).
Instead of using partitions 3 and 5, I prefer to
move the extended partition to slot 3, so that
the two rootfs partitions are in slot 5 and 6,
like they were when there was still a rescue
partition.