implement power-on from right button
This feature uses the wake-up mechanism from the Raspberry Pi
firmware. The firmware reboots into a "halt" state when the
boot partition ID indicated in the watchdog RSTS register is
set to 63. It can be woken up by a negative edge on GPIO3.
In order for this to be triggered by one of our buttons, we
change the pinout to connect the button to GPIO3 and connect
the display to I2C6 (it used to be on I2C1 which can no longer
be used).